+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [#1667] Implement `Mesh Cutter` - a component which can be used to delete or toggle portions of a mesh.
   - Implement vertex filters `By Axis` and `By Mask` (#1651) 
+- [#1697] Exposed `ModularAvatarMergeArmature.GetBonesMapping` API
 
 ### Fixed
 - [#1671] Shape changer could cause VRChat crashes in certain worlds